+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [3.0.2]
 - Fixed argument issues not highlighting the right location.
+- Improved indexing speed and skip indexing in non-workspace environments.
 
 ## [3.0.1]
 - Fixed extension breaking in non-workspace environments.